requirements.txt
flask
numpy
scipy
matplotlib
pillow